requirements.txt
absl-py==2.1.0
alembic==1.15.2
astunparse==1.6.3
audiomentations==0.39.0
audioread==3.0.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.1.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
gTTS==2.5.4
gunicorn==23.0.0
h5py==3.13.0
huggingface-hub==0.29.3
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
numpy-minmax==0.3.1
numpy-rms==0.4.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-stretch==0.2.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
soundfile==0.13.1
soxr==0.5.0.post1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.40
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf==1.0.3
tf_keras==2.19.0
threadpoolctl==3.5.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
xgboost==2.1.4
